+++ android/src/main/java/com/logicwind/reactnativestatusbarheight/ReactNativeStatusBarHeightModule.kt
@@ -1,9 +1,10 @@
 package com.logicwind.reactnativestatusbarheight
 
+import android.app.Activity
+import android.os.Build
+import android.view.WindowInsets
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactContextBaseJavaModule
-import com.facebook.react.bridge.ReactMethod
-import com.facebook.react.bridge.Promise
 
 class ReactNativeStatusBarHeightModule(reactContext: ReactApplicationContext) :
   ReactContextBaseJavaModule(reactContext) {
@@ -12,11 +13,35 @@ class ReactNativeStatusBarHeightModule(reactContext: ReactApplicationContext) :
     return NAME
   }
 
-  // Example method
-  // See https://reactnative.dev/docs/native-modules-android
-  @ReactMethod
-  fun multiply(a: Double, b: Double, promise: Promise) {
-    promise.resolve(a * b)
+  override fun getConstants(): Map<String, Any> {
+    val constants = mutableMapOf<String, Any>()
+
+    val activity: Activity? = currentActivity
+    var bottomInset = 0
+
+    if (activity != null) {
+      val rootView = activity.window.decorView.rootView
+      val insets: WindowInsets? = rootView.rootWindowInsets
+
+      if (insets != null) {
+        bottomInset = when {
+          Build.VERSION.SDK_INT >= Build.VERSION_CODES.R -> {
+            insets.getInsets(WindowInsets.Type.systemBars()).bottom
+          }
+          Build.VERSION.SDK_INT >= Build.VERSION_CODES.M -> {
+            insets.systemWindowInsetBottom
+          }
+          else -> 0
+        }
+      }
+    }
+
+    val density = reactApplicationContext.resources.displayMetrics.density
+    val bottomInsetDp = bottomInset / density
+
+    constants["BOTTOM_SAFE_AREA"] = bottomInsetDp
+
+    return constants
   }
 
   companion object {